Lobby methods moved to lobby controller and adjusted to be rest
