+++ src/main/kotlin/com/vakzu/musicwars/controllers/MainViewController.kt
@@ -1,10 +1,6 @@
 package com.vakzu.musicwars.controllers
 
 import com.vakzu.musicwars.dto.*
-import com.vakzu.musicwars.dto.websocket.CommandType
-import com.vakzu.musicwars.dto.websocket.OnlineMessage
-import com.vakzu.musicwars.dto.websocket.ReadyResponse
-import com.vakzu.musicwars.dto.websocket.SetReadyRequest
 import com.vakzu.musicwars.lobby.LobbyService
 import com.vakzu.musicwars.repos.CharacterRepository
 import com.vakzu.musicwars.repos.EffectRepository
@@ -12,8 +8,6 @@ import com.vakzu.musicwars.repos.SongRepository
 import com.vakzu.musicwars.security.MyUserPrincipal
 import com.vakzu.musicwars.security.UserService
 import com.vakzu.musicwars.services.FightService
-import org.springframework.http.HttpStatus
-import org.springframework.http.ResponseEntity
 import org.springframework.messaging.simp.SimpMessagingTemplate
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
 import org.springframework.stereotype.Controller
@@ -23,12 +17,10 @@ import java.security.Principal
 
 @Controller
 class MainViewController(
-    val messagingTemplate: SimpMessagingTemplate,
     val userService: UserService,
     val lobbyService: LobbyService,
     val effectRepository: EffectRepository,
     val characterRepository: CharacterRepository,
-    val fightService: FightService,
     val songRepository: SongRepository
 ) {
 
@@ -71,75 +63,5 @@ class MainViewController(
         return "game_page"
     }
 
-    @PostMapping("/lobby/create")
-    fun createLobby(principal: Principal): String {
-        val user = ((principal as UsernamePasswordAuthenticationToken).principal as MyUserPrincipal).user
-        val uuid = lobbyService.createLobby(user)
-        return "redirect:/war/$uuid"
-    }
-
-    @PostMapping("/lobby/join")
-    fun acceptInvite(@RequestParam("lobbyId") lobbyId: String, principal: Principal): String {
-        val user = ((principal as UsernamePasswordAuthenticationToken).principal as MyUserPrincipal).user
-        val lobby = lobbyService.getLobby(lobbyId)
-        lobby.addParticipant(user)
-
-        messagingTemplate.convertAndSend("/topic/lobby/$lobbyId", OnlineMessage(CommandType.JOIN, user.id, user.name))
-
-        return "redirect:/war/$lobbyId"
-    }
-
-    @PostMapping("/lobby/leave")
-    fun leaveLobby(@RequestParam("lobbyId") lobbyId: String, principal: Principal): String {
-        val user = ((principal as UsernamePasswordAuthenticationToken).principal as MyUserPrincipal).user
-        val lobby = lobbyService.getLobby(lobbyId)
-        lobby.removeParticipant(user)
-        if (lobby.participants.isEmpty()) {
-            lobbyService.lobbies.remove(lobby.lobbyId)
-        }
-
-        messagingTemplate.convertAndSend("/topic/lobby/$lobbyId", OnlineMessage(CommandType.LEAVE, user.id, user.name))
-
-        return "redirect:/"
-    }
-
-    @PostMapping("/war/{lobbyId}/ready/set")
-    @ResponseBody
-    fun setReady(@PathVariable lobbyId: String, @RequestBody(required = true) readyRequest: SetReadyRequest, principal: Principal) {
-        val user = ((principal as UsernamePasswordAuthenticationToken).principal as MyUserPrincipal).user
-        val lobby = lobbyService.getLobby(lobbyId)
-        lobby.setReady(user, readyRequest)
-        messagingTemplate.convertAndSend("/topic/lobby/$lobbyId", ReadyResponse(CommandType.SET_READY, user.id))
-    }
-
-    @PostMapping("/war/{lobbyId}/ready/cancel")
-    @ResponseBody
-    fun cancelReady(@PathVariable lobbyId: String, principal: Principal) {
-        val user = ((principal as UsernamePasswordAuthenticationToken).principal as MyUserPrincipal).user
-        val lobby = lobbyService.getLobby(lobbyId)
-        lobby.cancelReady(user)
-        messagingTemplate.convertAndSend("/topic/lobby/$lobbyId", ReadyResponse(CommandType.CANCEL_READY, user.id))
-    }
-
-    @PostMapping("/war/{lobbyId}/start")
-    @ResponseBody
-    fun beginFight(@PathVariable lobbyId: String, @RequestParam locationId: Int): ResponseEntity<*> {
-        val lobby = lobbyService.getLobby(lobbyId)
-        lobby.locationId = locationId
-        if (lobby.isEveryoneReady()) {
-            val moves =  fightService.playFight(lobby)
-            val fightMoves = moves.map { FightMoveResponse(it.moveNumber, it.fightId, it.attackerId, it.victimId, it.damage) }
-            messagingTemplate.convertAndSend("/topic/lobby/$lobbyId", fightMoves)
-            return ResponseEntity<Void>(HttpStatus.OK)
-        }
-        return ResponseEntity<Void>(HttpStatus.BAD_REQUEST)
-    }
-
-    @GetMapping("/characters/{characterId}/songs")
-    @ResponseBody
-    fun getAvailableSongs(@PathVariable characterId: Int): List<SongDto> {
-        val songs = songRepository.getCharacterAvailableSongs(characterId)
-        return songs.map { SongDto(it.id, it.name, it.damage) }
-    }
 
 }
